Task solution

+++ src/main/java/ru/kata/spring/boot_security/demo/repository/RoleRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.kata.spring.boot_security.demo.model.Role;
 
+import javax.persistence.criteria.CriteriaBuilder;
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+    //Role findByRoleType(String roleType);
 }
